+++ packages/providers/src/baseurl.tsx
@@ -1,0 +1,25 @@
+import React, { useContext } from 'react';
+
+const BaseUrlContext = React.createContext<{
+  baseurl?: string;
+}>({});
+
+export function BaseUrlProvider({
+  baseurl,
+  children,
+}: {
+  baseurl?: string;
+  children: React.ReactNode;
+}) {
+  return <BaseUrlContext.Provider value={{ baseurl }}>{children}</BaseUrlContext.Provider>;
+}
+
+export function useBaseurl() {
+  const data = useContext(BaseUrlContext);
+  return data?.baseurl;
+}
+
+export function withBaseurl(url?: string, baseurl?: string) {
+  if (baseurl) return baseurl + url;
+  return url as string;
+}