+++ packages/providers/src/references.tsx
@@ -5,22 +5,19 @@ import type { PageFrontmatter } from 'myst-frontmatter';
 const ReferencesContext = React.createContext<{
   frontmatter?: PageFrontmatter;
   references?: References;
-  urlbase?: string;
 }>({});
 
 export function ReferencesProvider({
   references,
   frontmatter,
-  urlbase,
   children,
 }: {
   frontmatter?: PageFrontmatter;
   references?: References;
-  urlbase?: string;
   children: React.ReactNode;
 }) {
   return (
-    <ReferencesContext.Provider value={{ references, frontmatter, urlbase }}>
+    <ReferencesContext.Provider value={{ references, frontmatter }}>
       {children}
     </ReferencesContext.Provider>
   );
@@ -34,14 +31,4 @@ export function useReferences() {
 export function useFrontmatter() {
   const data = useContext(ReferencesContext);
   return data?.frontmatter;
-}
-
-export function useUrlbase() {
-  const data = useContext(ReferencesContext);
-  return data?.urlbase;
-}
-
-export function withUrlbase(url?: string, urlbase?: string) {
-  if (urlbase) return urlbase + url;
-  return url as string;
 }